parameterize tokenlist
